changed order items

+++ db/migrate/20200826170351_create_orders.rb
@@ -2,8 +2,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
       t.integer :user_id
-      t.integer :total_price, default: 0
-
+      t.float :total_price, default: 0.00
       t.timestamps
     end
   end
